# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ARQON Revenue Engine V1 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizations:
    get:
      operationId: listOrganizations
      tags: [organizations]
      summary: List organizations for current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Organization"
    post:
      operationId: createOrganization
      tags: [organizations]
      summary: Create a new organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
  /organizations/{id}:
    get:
      operationId: getOrganization
      tags: [organizations]
      summary: Get organization by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateOrganization
      tags: [organizations]
      summary: Update organization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
    delete:
      operationId: deleteOrganization
      tags: [organizations]
      summary: Delete organization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content
  /organizations/{id}/stats:
    get:
      operationId: getOrganizationStats
      tags: [organizations]
      summary: Get revenue stats summary for an organization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationStats"
...
  /organizations/{orgId}/recommendations:
    get:
      operationId: getRecommendations
      tags: [recommendations]
      summary: List recommendations for an organization
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recommendation"
  /organizations/{orgId}/recommendations/{id}:
    get:
      operationId: getRecommendation
      tags: [recommendations]
      summary: Get recommendation by ID
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recommendation"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateRecommendation
      tags: [recommendations]
      summary: Update recommendation status
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecommendationUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recommendation"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    OrganizationInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        industry:
          type: string
        currency:
          type: string
          default: USD
    OrganizationUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        industry:
          type: string
        currency:
          type: string
    Organization:
      type: object
      required: [id, name, currency, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        industry:
          type: ["string", "null"]
        currency:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrganizationStats:
      type: object
      required: [totalAtRisk, findingsCount, openFindingsCount, resolvedFindingsCount, recommendationsCount, pendingRecommendationsCount, byType, bySeverity]
      properties:
        totalAtRisk:
          type: number
          description: Total estimated revenue at risk in organization currency
        findingsCount:
          type: integer
        openFindingsCount:
          type: integer
        resolvedFindingsCount:
          type: integer
        recommendationsCount:
          type: integer
        pendingRecommendationsCount:
          type: integer
        byType:
          type: array
          items:
            $ref: "#/components/schemas/FindingTypeStat"
        bySeverity:
          type: array
          items:
            $ref: "#/components/schemas/FindingSeverityStat"
    FindingTypeStat:
      type: object
      required: [type, count, totalAtRisk]
      properties:
        type:
          type: string
        count:
          type: integer
        totalAtRisk:
          type: number
    FindingSeverityStat:
      type: object
      required: [severity, count, totalAtRisk]
      properties:
        severity:
          type: string
        count:
          type: integer
        totalAtRisk:
          type: number
...
    Recommendation:
      type: object
      required: [id, orgId, findingId, priority, title, description, estimatedRecovery, status, createdAt]
      properties:
        id:
          type: integer
        orgId:
          type: integer
        findingId:
          type: integer
        priority:
          type: string
          enum: [critical, high, medium, low]
        title:
          type: string
        description:
          type: string
        estimatedRecovery:
          type: number
          description: Estimated recoverable revenue
        actionLabel:
          type: ["string", "null"]
          description: Short action button label
        status:
          type: string
          enum: [pending, in_progress, completed, dismissed]
        completedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RecommendationUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, in_progress, completed, dismissed]
...