# lib/api-zod/src/generated/types/findingTypeStat.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ARQON Revenue Engine V1 API
 * OpenAPI spec version: 0.1.0
 */

export interface FindingTypeStat {
  type: string;
  count: number;
  totalAtRisk: number;
}