# lib/api-zod/src/generated/types/organizationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ARQON Revenue Engine V1 API
 * OpenAPI spec version: 0.1.0
 */

export interface OrganizationInput {
  /** @minLength 1 */
  name: string;
  industry?: string;
  currency?: string;
}