# lib/api-zod/src/generated/types/organizationStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ARQON Revenue Engine V1 API
 * OpenAPI spec version: 0.1.0
 */
import type { FindingSeverityStat } from './findingSeverityStat';
import type { FindingTypeStat } from './findingTypeStat';

export interface OrganizationStats {
  /** Total estimated revenue at risk in organization currency */
  totalAtRisk: number;
  findingsCount: number;
  openFindingsCount: number;
  resolvedFindingsCount: number;
  recommendationsCount: number;
  pendingRecommendationsCount: number;
  byType: FindingTypeStat[];
  bySeverity: FindingSeverityStat[];
}