# lib/api-zod/src/generated/types/recommendation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ARQON Revenue Engine V1 API
 * OpenAPI spec version: 0.1.0
 */
import type { RecommendationPriority } from './recommendationPriority';
import type { RecommendationStatus } from './recommendationStatus';

export interface Recommendation {
  id: number;
  orgId: number;
  findingId: number;
  priority: RecommendationPriority;
  title: string;
  description: string;
  /** Estimated recoverable revenue */
  estimatedRecovery: number;
  /**
     * Short action button label
     * @nullable
     */
  actionLabel?: string | null;
  status: RecommendationStatus;
  /** @nullable */
  completedAt?: Date | null;
  createdAt: Date;
  updatedAt?: Date;
}